Refactor to reduce duplication
