+++ svelte/src/lib/index.js
@@ -1,21 +1,19 @@
 import questions_raw from "../assets/questions.txt?raw";
 import survey_questions_raw from "../assets/survey-questions.txt?raw";
 
-export const questionsUnrandomized = questions_raw
-  .split("\n")
-  .filter((e) => e) // filter out any blank questions (blank lines in questions.txt)
-  .map((question_text) => {
-    return { question_text, hash: hash(question_text) };
-  })
-  .sort((a, b) => a.question_text.localeCompare(b.question_text));
+function processQuestions(rawQuestions) {
+  return rawQuestions
+    .split("\n")
+    .filter((e) => e) // filter out any blank questions (blank lines in questions.txt)
+    .map((question_text) => {
+      return { question_text, hash: hash(question_text) };
+    })
+    .sort((a, b) => a.question_text.localeCompare(b.question_text));
+}
 
-export const surveyQuestionsUnrandomized = survey_questions_raw
-  .split("\n")
-  .filter((e) => e) // filter out any blank questions (blank lines in questions.txt)
-  .map((question_text) => {
-    return { question_text, hash: hash(question_text) };
-  })
-  .sort((a, b) => a.question_text.localeCompare(b.question_text));
+export const questionsUnrandomized = processQuestions(questions_raw);
+
+export const surveyQuestionsUnrandomized = processQuestions(survey_questions_raw);
 
 export function hash(s) {
   const hash = Array.from(s).reduce((hash, char) => {
