+++ svelte/src/routes/+layout.svelte
@@ -1,5 +1,4 @@
 <script>
-    export const prerender = true;
 </script>
 
 <slot />
